# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api sistema de agendacion de citas asopormen",
        "description": "Api para la gestion de citas a usuarios de la ips Asopormen.sp",
        "termsOfService": "https://asopormen.org.co/",
        "contact": {
            "name": "Equipo de soporte",
            "email": "msolegario.cisneros@gmail.com"
        },
        "license": {
            "name": "Licencia",
            "url": "https://example.com/license"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://por_definir/api/documentation"
        }
    ],
    "paths": {
...
        "/api/get_clients/{string}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Buscar clientes por nombre",
                "description": "Busca clientes cuyos nombres contengan la cadena proporcionada. Solo se devuelven los clientes activos y cuyo `ok_ent` es 0. Limita el número de resultados a 20.",
                "operationId": "c594718b34cc1475fe8e7fda382959c2",
                "parameters": [
                    {
                        "name": "string",
                        "in": "path",
                        "description": "Cadena para buscar en el nombre de los clientes",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "John"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clientes encontrados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "codigo": {
                                                        "type": "string",
                                                        "example": "123"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetro de búsqueda no válido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "parametro de búsqueda debe ser válido"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Clientes no encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "no se encontraron clientes"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/get_client": {
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener información del cliente por historyId",
                "description": "Obtiene la información de un cliente usando un `historyId` numérico proporcionado en la solicitud. La respuesta incluye detalles del cliente como nombre, nit, fecha de nacimiento, entre otros.",
                "operationId": "ff9071d0cf0ce9d2571ede590fa11bba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "historyId": {
                                        "description": "ID histórico del cliente. solo caracteres numericos.",
                                        "type": "string",
                                        "example": 123456
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Información del cliente obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "codigo": {
                                                        "type": "string",
                                                        "example": "12345"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "nit_cli": {
                                                        "type": "string",
                                                        "example": "900123456"
                                                    },
                                                    "f_nacio": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "1990-01-01"
                                                    },
                                                    "sexo": {
                                                        "type": "string",
                                                        "example": "M"
                                                    },
                                                    "direcc": {
                                                        "type": "string",
                                                        "example": "123 Main St"
                                                    },
                                                    "barrio": {
                                                        "type": "string",
                                                        "example": "Downtown"
                                                    },
                                                    "cel": {
                                                        "type": "string",
                                                        "example": "555-1234"
                                                    },
                                                    "municipio": {
                                                        "type": "string",
                                                        "example": "CityName"
                                                    },
                                                    "entidad": {
                                                        "type": "string",
                                                        "example": "EntityName"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetro `historyId` inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "parametro historyId debe ser válido"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Información del cliente no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "informacion de usuario no encontrada"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}